Phase 4 batch 1: extract FeedbackPanel out of SubmissionStatusView
SubmissionStatusView.tsx was 2,602 lines, the largest .tsx in the app. The
revision-composer component (builder switching, attachments, sketch upload,
send/stop, quick actions) was already a self-contained function taking only
props and its own local state, so it moves out whole into its own file with
no behavior change: SubmissionStatusView.tsx drops to 1,956 lines, and
FeedbackPanel.tsx is a new 654-line file with its own module-size and
comment-prose baseline.

All 83 SubmissionStatusView tests and the full 1,889-test web suite pass
unchanged; the workspace build and type-check are clean.
